+++ test/functional/replset_operations_tests.js
@@ -37,7 +37,7 @@ exports['Should fail due to w:5 and wtimeout:1 with ordered batch api'] = {
     // Create url
     var url = format("mongodb://%s,%s/%s?replicaSet=%s&readPreference=%s"
       , format("%s:%s", configuration.host, configuration.port)
-      , format("%s:%s", configuration.host, configuration.host + 1)
+      , format("%s:%s", configuration.host, configuration.port + 1)
       , "integration_test_"
       , configuration.replicasetName
       , "primary");
@@ -83,6 +83,7 @@ exports['Should fail due to w:5 and wtimeout:1 with ordered batch api'] = {
     }
 
     MongoClient.connect(url, function(err, db) {
+      console.dir(err)
       test.equal(null, err);
 
       executeTests(db, function() {
@@ -357,7 +358,7 @@ exports['Should Correctly group using replicaset'] = {
     // Create url
     var url = format("mongodb://%s,%s/%s?replicaSet=%s&readPreference=%s"
       , format("%s:%s", configuration.host, configuration.port)
-      , format("%s:%s", configuration.host, configuration.host + 1)
+      , format("%s:%s", configuration.host, configuration.port + 1)
       , "integration_test_"
       , configuration.replicasetName
       , "primary");
@@ -402,7 +403,7 @@ exports['Should fail to do map reduce to out collection'] = {
     // Create url
     var url = format("mongodb://%s,%s/%s?replicaSet=%s&readPreference=%s"
       , format("%s:%s", configuration.host, configuration.port)
-      , format("%s:%s", configuration.host, configuration.host + 1)
+      , format("%s:%s", configuration.host, configuration.port + 1)
       , "integration_test_"
       , configuration.replicasetName
       , "primary");