+++ HISTORY.md
@@ -2,6 +2,7 @@
 -----------------
 * Propagate timeout event correctly to db instances.
 * Application Monitoring API (APM) implemented.
+* NOT providing replSet name in MongoClient connection URI will force single server connection. Fixes issue where it was impossible to directly connect to a replicaset member server.
 
 2.0.42 08-18-2015
 -----------------